+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Update zsh shell completion (#264)
+
 ## [1.7.0] - 2019-03-25
 
 ### Added